+++ RELEASE_NOTES.md
@@ -9,6 +9,7 @@
 * Improvements to Mp3FileReader seeking and position reporting (thanks to protyposis)
 * updated NAudio build process to use FAKE, retiring the old MSBuild and IronPython scripts
 * NAudio.Wma project is moved out into its own [GitHub repository](https://github.com/naudio/NAudio.Wma)
+* ConcatenatingSampleProvider and FollowedBy extension method making it easy to concatenate Sample Providers
 * Various bugfixes and enhancements. See commit log for full details
   * WasapiCapture buffer sizes can be specified 
   * MMDeviceEnumerator is disposable